requirements.txt
Babel==2.6.0
boto3==1.9.71
botocore==1.12.71
certifi==2018.10.15
chardet==3.0.4
Click==7.0
dj-database-url==0.5.0
Django==2.1
django-extra-views==0.11.0
django-haystack==2.8.1
django-heroku==0.3.1
django-phonenumber-field==2.0.1
django-storages==1.7.1
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.3
docutils==0.14
factory-boy==2.11.1
Faker==1.0.0
Flask==1.0.2
gunicorn==19.9.0
idna==2.7
image==1.5.27
ItsDangerous==1.0.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
mock==2.0.0
pbr==5.1.0
phonenumbers==8.10.1
Pillow==5.3.0
pipenv==2018.10.13
psycopg2==2.7.6.1
purl==1.4
pypi==2.1
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.0
s3transfer==0.1.13
simplejson==3.16.0
six==1.11.0
sorl-thumbnail==12.4.1
stevedore==1.30.0
text-unidecode==1.2
Unidecode==1.0.23
urllib3==1.24
virtualenv==16.0.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
web.py==0.40.dev1
Werkzeug==0.14.1
whitenoise==4.1.1
